requirements.txt
absl-py==2.3.0
Adafruit-Blinka==8.61.2
adafruit-circuitpython-74hc595==1.4.8
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-charlcd==3.5.3
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-mcp230xx==2.5.18
adafruit-circuitpython-pca9685==3.4.19
adafruit-circuitpython-register==1.10.4
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.80.0
Adafruit-PureIO==1.1.11
astunparse==1.6.3
binho-host-adapter==0.1.6
certifi==2025.6.15
charset-normalizer==3.4.2
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.5
pyclipper==1.3.0.post6
pyftdi==0.56.0
Pygments==2.19.2
pyparsing==3.2.3
pypiwin32==223
pyserial==3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pyttsx3==2.98
pyusb==1.3.1
pywin32==310
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scipy==1.16.0
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
